+++ lib/models/objects.dart
@@ -1,4 +1,0 @@
-enum NotionObjects {
-  user,
-  database,
-}