+++ lib/models/rich_text/notion_types.dart
@@ -1,0 +1,6 @@
+enum NotionTypes {
+  H1,
+  H2,
+  H3,
+  Paragraph,
+}